refactor(code-splitting): extract the simulated-facade inclusion replay into shared helpers
